ENH: get rid of the defaults file
The defaults file was used to define global settings. This commit drops
the defaults file as a separate files and instead uses the 'global' key
in the runtime conf to define settings that should be effective in every
bot.
To accomplish that there is now a utils.get_global_settings() function
that extracts the 'global' dict from the runtime configuration file.